Tournament games to NN input
Wrote a converted that takes the tournament games (found at http://www.ffothello.org/informatique/la-base-wthor/) and converts them to input for the Neural Net. The input for the neural net is the state of the board, the acting player, the coordinates of the next move (for policy net) and the marginal score of the game (for the value net). Loads of thanks to Ledpup (http://ledpup.blogspot.nl/2012/03/computer-reversi-part-15-thor-database.html) for getting me started on parsing the tournament files.